tests: Add Push Descriptor test files
