+++ tests/unit/command.cpp
@@ -4595,146 +4595,6 @@ TEST_F(NegativeCommand, SecondaryCommandBufferClearColorAttachmentsRenderArea) {
     m_commandBuffer->end();
 }
 
-TEST_F(NegativeCommand, PushDescriptorSetCmdPush) {
-    TEST_DESCRIPTION("Attempt to push a push descriptor set with incorrect arguments.");
-    AddRequiredExtensions(VK_KHR_PUSH_DESCRIPTOR_EXTENSION_NAME);
-
-    ASSERT_NO_FATAL_FAILURE(InitFramework(m_errorMonitor));
-    if (!AreRequiredExtensionsEnabled()) {
-        GTEST_SKIP() << RequiredExtensionsNotSupported() << " not supported";
-    }
-    ASSERT_NO_FATAL_FAILURE(InitState());
-
-    // Create ordinary and push descriptor set layout
-    VkDescriptorSetLayoutBinding binding = {0, VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER, 1, VK_SHADER_STAGE_FRAGMENT_BIT, nullptr};
-    const VkDescriptorSetLayoutObj ds_layout(m_device, {binding});
-    ASSERT_TRUE(ds_layout.initialized());
-    const VkDescriptorSetLayoutObj push_ds_layout(m_device, {binding}, VK_DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR);
-    ASSERT_TRUE(push_ds_layout.initialized());
-
-    // Now use the descriptor set layouts to create a pipeline layout
-    const VkPipelineLayoutObj pipeline_layout(m_device, {&push_ds_layout, &ds_layout});
-    ASSERT_TRUE(pipeline_layout.initialized());
-
-    // Create a descriptor to push
-    const uint32_t buffer_data[4] = {4, 5, 6, 7};
-    VkConstantBufferObj buffer_obj(m_device, sizeof(buffer_data), &buffer_data);
-    ASSERT_TRUE(buffer_obj.initialized());
-
-    // Create a "write" struct, noting that the buffer_info cannot be a temporary arg (the return from write_descriptor_set
-    // references its data), and the DescriptorSet() can be temporary, because the value is ignored
-    VkDescriptorBufferInfo buffer_info = {buffer_obj.handle(), 0, VK_WHOLE_SIZE};
-
-    VkWriteDescriptorSet descriptor_write = vk_testing::Device::write_descriptor_set(
-        vk_testing::DescriptorSet(), 0, 0, VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER, 1, &buffer_info);
-
-    // Section 1: Queue family matching/capabilities.
-    // Create command pool on a non-graphics queue
-    const std::optional<uint32_t> no_gfx_qfi = m_device->QueueFamilyMatching(VK_QUEUE_COMPUTE_BIT, VK_QUEUE_GRAPHICS_BIT);
-    const std::optional<uint32_t> transfer_only_qfi =
-        m_device->QueueFamilyMatching(VK_QUEUE_TRANSFER_BIT, (VK_QUEUE_COMPUTE_BIT | VK_QUEUE_GRAPHICS_BIT));
-    if (transfer_only_qfi || no_gfx_qfi) {
-        const uint32_t err_qfi = no_gfx_qfi ? no_gfx_qfi.value() : transfer_only_qfi.value();
-
-        VkCommandPoolObj command_pool(m_device, err_qfi);
-        ASSERT_TRUE(command_pool.initialized());
-        VkCommandBufferObj command_buffer(m_device, &command_pool);
-        ASSERT_TRUE(command_buffer.initialized());
-        command_buffer.begin();
-
-        m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkCmdPushDescriptorSetKHR-pipelineBindPoint-00363");
-        m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkWriteDescriptorSet-descriptorType-00330");
-        if (err_qfi == transfer_only_qfi) {
-            // This as this queue neither supports the gfx or compute bindpoints, we'll get two errors
-            m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkCmdPushDescriptorSetKHR-commandBuffer-cmdpool");
-        }
-        vk::CmdPushDescriptorSetKHR(command_buffer.handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipeline_layout.handle(), 0, 1,
-                                    &descriptor_write);
-        m_errorMonitor->VerifyFound();
-        command_buffer.end();
-
-        // If we succeed in testing only one condition above, we need to test the other below.
-        if (transfer_only_qfi && err_qfi != transfer_only_qfi.value()) {
-            // Need to test the neither compute/gfx supported case separately.
-            VkCommandPoolObj tran_command_pool(m_device, transfer_only_qfi.value());
-            ASSERT_TRUE(tran_command_pool.initialized());
-            VkCommandBufferObj tran_command_buffer(m_device, &tran_command_pool);
-            ASSERT_TRUE(tran_command_buffer.initialized());
-            tran_command_buffer.begin();
-
-            // We can't avoid getting *both* errors in this case
-            m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkCmdPushDescriptorSetKHR-pipelineBindPoint-00363");
-            m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkWriteDescriptorSet-descriptorType-00330");
-            m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkCmdPushDescriptorSetKHR-commandBuffer-cmdpool");
-            vk::CmdPushDescriptorSetKHR(tran_command_buffer.handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipeline_layout.handle(), 0,
-                                        1, &descriptor_write);
-            m_errorMonitor->VerifyFound();
-            tran_command_buffer.end();
-        }
-    }
-
-    // Push to the non-push binding
-    m_commandBuffer->begin();
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkCmdPushDescriptorSetKHR-set-00365");
-    vk::CmdPushDescriptorSetKHR(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipeline_layout.handle(), 1, 1,
-                                &descriptor_write);
-    m_errorMonitor->VerifyFound();
-
-    // Specify set out of bounds
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkCmdPushDescriptorSetKHR-set-00364");
-    vk::CmdPushDescriptorSetKHR(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipeline_layout.handle(), 2, 1,
-                                &descriptor_write);
-    m_errorMonitor->VerifyFound();
-    m_commandBuffer->end();
-
-    // TODO: Add VALIDATION_ERROR_ code support to core_validation::ValidateCmd
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkCmdPushDescriptorSetKHR-commandBuffer-recording");
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkWriteDescriptorSet-descriptorType-00330");
-    vk::CmdPushDescriptorSetKHR(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipeline_layout.handle(), 0, 1,
-                                &descriptor_write);
-    m_errorMonitor->VerifyFound();
-}
-
-TEST_F(NegativeCommand, PushDescriptorSetCmdBufferOffsetUnaligned) {
-    TEST_DESCRIPTION("Attempt to push a push descriptor set buffer with unaligned offset.");
-    AddRequiredExtensions(VK_KHR_PUSH_DESCRIPTOR_EXTENSION_NAME);
-
-    ASSERT_NO_FATAL_FAILURE(InitFramework(m_errorMonitor));
-    if (!AreRequiredExtensionsEnabled()) {
-        GTEST_SKIP() << RequiredExtensionsNotSupported() << " not supported";
-    }
-    ASSERT_NO_FATAL_FAILURE(InitState());
-
-    auto const min_alignment = m_device->props.limits.minUniformBufferOffsetAlignment;
-    if (min_alignment == 0) {
-        GTEST_SKIP() << "minUniformBufferOffsetAlignment is zero";
-    }
-
-    VkDescriptorSetLayoutBinding binding = {0, VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER, 1, VK_SHADER_STAGE_FRAGMENT_BIT, nullptr};
-    const VkDescriptorSetLayoutObj push_ds_layout(m_device, {binding}, VK_DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR);
-    ASSERT_TRUE(push_ds_layout.initialized());
-
-    const VkPipelineLayoutObj pipeline_layout(m_device, {&push_ds_layout});
-    ASSERT_TRUE(pipeline_layout.initialized());
-
-    const uint32_t buffer_data[4] = {4, 5, 6, 7};
-    VkConstantBufferObj buffer_obj(m_device, sizeof(buffer_data), &buffer_data, VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT);
-    ASSERT_TRUE(buffer_obj.initialized());
-
-    // Use an invalid alignment.
-    VkDescriptorBufferInfo buffer_info = {buffer_obj.handle(), min_alignment - 1, VK_WHOLE_SIZE};
-    VkWriteDescriptorSet descriptor_write = vk_testing::Device::write_descriptor_set(
-        vk_testing::DescriptorSet(), 0, 0, VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER, 1, &buffer_info);
-
-    m_commandBuffer->begin();
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkWriteDescriptorSet-descriptorType-00327");
-    vk::CmdPushDescriptorSetKHR(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipeline_layout.handle(), 0, 1,
-                                &descriptor_write);
-    m_errorMonitor->VerifyFound();
-
-    m_commandBuffer->end();
-}
-
 TEST_F(NegativeCommand, MultiDraw) {
     TEST_DESCRIPTION("Test validation of multi_draw extension");
     SetTargetApiVersion(VK_API_VERSION_1_2);
