+++ tests/unit/descriptors.cpp
@@ -1953,132 +1953,6 @@ TEST_F(NegativeDescriptors, DSUsageBitsFlags2) {
     m_errorMonitor->VerifyFound();
 }
 
-TEST_F(NegativeDescriptors, DSBufferInfo) {
-    TEST_DESCRIPTION(
-        "Attempt to update buffer descriptor set that has incorrect parameters in VkDescriptorBufferInfo struct. This includes:\n"
-        "1. offset value greater than or equal to buffer size\n"
-        "2. range value of 0\n"
-        "3. range value greater than buffer (size - offset)");
-
-    AddRequiredExtensions(VK_KHR_DESCRIPTOR_UPDATE_TEMPLATE_EXTENSION_NAME);
-    AddRequiredExtensions(VK_KHR_PUSH_DESCRIPTOR_EXTENSION_NAME);
-    ASSERT_NO_FATAL_FAILURE(InitFramework());
-    if (!AreRequiredExtensionsEnabled()) {
-        GTEST_SKIP() << RequiredExtensionsNotSupported() << " not supported.";
-    }
-    ASSERT_NO_FATAL_FAILURE(InitState(nullptr, nullptr, VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT));
-
-    std::vector<VkDescriptorSetLayoutBinding> ds_bindings = {
-        {0, VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER, 1, VK_SHADER_STAGE_ALL, nullptr}};
-    OneOffDescriptorSet descriptor_set(m_device, ds_bindings);
-
-    // Create a buffer to be used for invalid updates
-    auto buff_ci = LvlInitStruct<VkBufferCreateInfo>();
-    buff_ci.usage = VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT;
-    buff_ci.size = m_device->props.limits.minUniformBufferOffsetAlignment;
-    buff_ci.sharingMode = VK_SHARING_MODE_EXCLUSIVE;
-    VkBufferObj buffer(*m_device, buff_ci);
-
-    VkDescriptorBufferInfo buff_info = {};
-    buff_info.buffer = buffer.handle();
-    auto descriptor_write = LvlInitStruct<VkWriteDescriptorSet>();
-    descriptor_write.dstBinding = 0;
-    descriptor_write.descriptorCount = 1;
-    descriptor_write.pTexelBufferView = nullptr;
-    descriptor_write.pBufferInfo = &buff_info;
-    descriptor_write.pImageInfo = nullptr;
-
-    descriptor_write.descriptorType = VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER;
-    descriptor_write.dstSet = descriptor_set.set_;
-
-    // Setup for update w/ template tests
-    // Create a template of descriptor set updates
-    struct SimpleTemplateData {
-        uint8_t padding[7];
-        VkDescriptorBufferInfo buff_info;
-        uint32_t other_padding[4];
-    };
-    SimpleTemplateData update_template_data = {};
-
-    VkDescriptorUpdateTemplateEntry update_template_entry = {};
-    update_template_entry.dstBinding = 0;
-    update_template_entry.dstArrayElement = 0;
-    update_template_entry.descriptorCount = 1;
-    update_template_entry.descriptorType = VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER;
-    update_template_entry.offset = offsetof(SimpleTemplateData, buff_info);
-    update_template_entry.stride = sizeof(SimpleTemplateData);
-
-    auto update_template_ci = LvlInitStruct<VkDescriptorUpdateTemplateCreateInfoKHR>();
-    update_template_ci.descriptorUpdateEntryCount = 1;
-    update_template_ci.pDescriptorUpdateEntries = &update_template_entry;
-    update_template_ci.templateType = VK_DESCRIPTOR_UPDATE_TEMPLATE_TYPE_DESCRIPTOR_SET;
-    update_template_ci.descriptorSetLayout = descriptor_set.layout_.handle();
-
-    VkDescriptorUpdateTemplate update_template = VK_NULL_HANDLE;
-    ASSERT_VK_SUCCESS(vk::CreateDescriptorUpdateTemplateKHR(m_device->device(), &update_template_ci, nullptr, &update_template));
-
-    std::unique_ptr<VkDescriptorSetLayoutObj> push_dsl = nullptr;
-    std::unique_ptr<VkPipelineLayoutObj> pipeline_layout = nullptr;
-    VkDescriptorUpdateTemplate push_template = VK_NULL_HANDLE;
-
-    push_dsl.reset(new VkDescriptorSetLayoutObj(m_device, ds_bindings, VK_DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR));
-    pipeline_layout.reset(new VkPipelineLayoutObj(m_device, {push_dsl.get()}));
-    ASSERT_TRUE(push_dsl->initialized());
-
-    auto push_template_ci = LvlInitStruct<VkDescriptorUpdateTemplateCreateInfoKHR>();
-    push_template_ci.descriptorUpdateEntryCount = 1;
-    push_template_ci.pDescriptorUpdateEntries = &update_template_entry;
-    push_template_ci.templateType = VK_DESCRIPTOR_UPDATE_TEMPLATE_TYPE_PUSH_DESCRIPTORS_KHR;
-    push_template_ci.descriptorSetLayout = VK_NULL_HANDLE;
-    push_template_ci.pipelineBindPoint = VK_PIPELINE_BIND_POINT_GRAPHICS;
-    push_template_ci.pipelineLayout = pipeline_layout->handle();
-    push_template_ci.set = 0;
-    ASSERT_VK_SUCCESS(vk::CreateDescriptorUpdateTemplateKHR(m_device->device(), &push_template_ci, nullptr, &push_template));
-
-    auto do_test = [&](const char *desired_failure) {
-        m_errorMonitor->SetDesiredFailureMsg(kErrorBit, desired_failure);
-        vk::UpdateDescriptorSets(m_device->device(), 1, &descriptor_write, 0, NULL);
-        m_errorMonitor->VerifyFound();
-
-        m_errorMonitor->SetDesiredFailureMsg(kErrorBit, desired_failure);
-        m_commandBuffer->begin();
-        vk::CmdPushDescriptorSetKHR(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipeline_layout->handle(), 0, 1,
-                                    &descriptor_write);
-        m_commandBuffer->end();
-        m_errorMonitor->VerifyFound();
-
-        update_template_data.buff_info = buff_info;  // copy the test case information into our "pData"
-        m_errorMonitor->SetDesiredFailureMsg(kErrorBit, desired_failure);
-        vk::UpdateDescriptorSetWithTemplateKHR(m_device->device(), descriptor_set.set_, update_template, &update_template_data);
-        m_errorMonitor->VerifyFound();
-
-        m_errorMonitor->SetDesiredFailureMsg(kErrorBit, desired_failure);
-        m_commandBuffer->begin();
-        vk::CmdPushDescriptorSetWithTemplateKHR(m_commandBuffer->handle(), push_template, pipeline_layout->handle(), 0,
-                                                &update_template_data);
-        m_commandBuffer->end();
-        m_errorMonitor->VerifyFound();
-    };
-
-    // Cause error due to offset out of range
-    buff_info.offset = buff_ci.size;
-    buff_info.range = VK_WHOLE_SIZE;
-    do_test("VUID-VkDescriptorBufferInfo-offset-00340");
-
-    // Now cause error due to range of 0
-    buff_info.offset = 0;
-    buff_info.range = 0;
-    do_test("VUID-VkDescriptorBufferInfo-range-00341");
-
-    // Now cause error due to range exceeding buffer size - offset
-    buff_info.offset = 0;
-    buff_info.range = buff_ci.size + 1;
-    do_test("VUID-VkDescriptorBufferInfo-range-00342");
-
-    vk::DestroyDescriptorUpdateTemplateKHR(m_device->device(), update_template, nullptr);
-    vk::DestroyDescriptorUpdateTemplateKHR(m_device->device(), push_template, nullptr);
-}
-
 TEST_F(NegativeDescriptors, DSBufferLimit) {
     TEST_DESCRIPTION(
         "Attempt to update buffer descriptor set that has VkDescriptorBufferInfo values that violate device limits.\n"
@@ -2606,120 +2480,6 @@ TEST_F(NegativeDescriptors, WriteDescriptorSetNotAllocated) {
     m_errorMonitor->VerifyFound();
 }
 
-TEST_F(NegativeDescriptors, PushDescriptorDestroyDescriptorSetLayout) {
-    TEST_DESCRIPTION("Delete the DescriptorSetLayout and then call vkCmdPushDescriptorSetKHR");
-
-    AddRequiredExtensions(VK_KHR_PUSH_DESCRIPTOR_EXTENSION_NAME);
-    ASSERT_NO_FATAL_FAILURE(InitFramework());
-    if (!AreRequiredExtensionsEnabled()) {
-        GTEST_SKIP() << RequiredExtensionsNotSupported() << " not supported";
-    }
-    ASSERT_NO_FATAL_FAILURE(InitState());
-
-    VkBufferObj buffer(*m_device, 32, VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT);
-
-    VkDescriptorSetLayoutBinding ds_binding = {0, VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER, 1, VK_SHADER_STAGE_ALL, nullptr};
-    VkDescriptorSetLayout ds_layout = VK_NULL_HANDLE;
-    auto dsl_ci = LvlInitStruct<VkDescriptorSetLayoutCreateInfo>();
-    dsl_ci.flags = VK_DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR;
-    dsl_ci.bindingCount = 1;
-    dsl_ci.pBindings = &ds_binding;
-    vk::CreateDescriptorSetLayout(m_device->device(), &dsl_ci, nullptr, &ds_layout);
-
-    auto pipeline_layout_ci = LvlInitStruct<VkPipelineLayoutCreateInfo>();
-    pipeline_layout_ci.setLayoutCount = 1;
-    pipeline_layout_ci.pSetLayouts = &ds_layout;
-    pipeline_layout_ci.pushConstantRangeCount = 0;
-    vk_testing::PipelineLayout pipeline_layout(*m_device, pipeline_layout_ci);
-
-    VkDescriptorBufferInfo buffer_info = {buffer.handle(), 0, 32};
-    auto descriptor_write = LvlInitStruct<VkWriteDescriptorSet>();
-    descriptor_write.dstBinding = 0;
-    descriptor_write.descriptorCount = 1;
-    descriptor_write.pTexelBufferView = nullptr;
-    descriptor_write.pBufferInfo = &buffer_info;
-    descriptor_write.pImageInfo = nullptr;
-    descriptor_write.descriptorType = VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER;
-
-    m_commandBuffer->begin();
-
-    vk::DestroyDescriptorSetLayout(m_device->device(), ds_layout, nullptr);
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkWriteDescriptorSet-dstSet-00320");
-    vk::CmdPushDescriptorSetKHR(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipeline_layout.handle(), 0, 1,
-                                &descriptor_write);
-    m_errorMonitor->VerifyFound();
-
-    m_commandBuffer->end();
-}
-
-TEST_F(NegativeDescriptors, PushDescriptorTemplateDestroyDescriptorSetLayout) {
-    TEST_DESCRIPTION("Delete the DescriptorSetLayout and then call vkCmdPushDescriptorSetWithTemplateKHR");
-
-    AddRequiredExtensions(VK_KHR_PUSH_DESCRIPTOR_EXTENSION_NAME);
-    AddRequiredExtensions(VK_KHR_DESCRIPTOR_UPDATE_TEMPLATE_EXTENSION_NAME);
-    ASSERT_NO_FATAL_FAILURE(InitFramework());
-    if (!AreRequiredExtensionsEnabled()) {
-        GTEST_SKIP() << RequiredExtensionsNotSupported() << " not supported";
-    }
-    ASSERT_NO_FATAL_FAILURE(InitState());
-
-    auto buffer_ci = LvlInitStruct<VkBufferCreateInfo>();
-    buffer_ci.size = 32;
-    buffer_ci.usage = VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT;
-    VkBufferObj buffer(*m_device, buffer_ci);
-
-    VkDescriptorSetLayoutBinding ds_binding = {0, VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER, 1, VK_SHADER_STAGE_ALL, nullptr};
-    VkDescriptorSetLayout ds_layout = VK_NULL_HANDLE;
-    auto dsl_ci = LvlInitStruct<VkDescriptorSetLayoutCreateInfo>();
-    dsl_ci.flags = VK_DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR;
-    dsl_ci.bindingCount = 1;
-    dsl_ci.pBindings = &ds_binding;
-    vk::CreateDescriptorSetLayout(m_device->device(), &dsl_ci, nullptr, &ds_layout);
-
-    auto pipeline_layout_ci = LvlInitStruct<VkPipelineLayoutCreateInfo>();
-    pipeline_layout_ci.setLayoutCount = 1;
-    pipeline_layout_ci.pSetLayouts = &ds_layout;
-    pipeline_layout_ci.pushConstantRangeCount = 0;
-    VkPipelineLayout pipeline_layout = VK_NULL_HANDLE;
-    vk::CreatePipelineLayout(m_device->device(), &pipeline_layout_ci, nullptr, &pipeline_layout);
-
-    struct SimpleTemplateData {
-        VkDescriptorBufferInfo buff_info;
-    };
-
-    VkDescriptorUpdateTemplateEntry update_template_entry = {};
-    update_template_entry.dstBinding = 0;
-    update_template_entry.dstArrayElement = 0;
-    update_template_entry.descriptorCount = 1;
-    update_template_entry.descriptorType = VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER;
-    update_template_entry.offset = offsetof(SimpleTemplateData, buff_info);
-    update_template_entry.stride = sizeof(SimpleTemplateData);
-
-    auto update_template_ci = LvlInitStruct<VkDescriptorUpdateTemplateCreateInfoKHR>();
-    update_template_ci.descriptorUpdateEntryCount = 1;
-    update_template_ci.pDescriptorUpdateEntries = &update_template_entry;
-    update_template_ci.templateType = VK_DESCRIPTOR_UPDATE_TEMPLATE_TYPE_PUSH_DESCRIPTORS_KHR;
-    update_template_ci.descriptorSetLayout = ds_layout;
-    update_template_ci.pipelineBindPoint = VK_PIPELINE_BIND_POINT_GRAPHICS;
-    update_template_ci.pipelineLayout = pipeline_layout;
-
-    VkDescriptorUpdateTemplate update_template = VK_NULL_HANDLE;
-    vk::CreateDescriptorUpdateTemplateKHR(m_device->device(), &update_template_ci, nullptr, &update_template);
-
-    SimpleTemplateData update_template_data;
-    update_template_data.buff_info = {buffer.handle(), 0, 32};
-
-    m_commandBuffer->begin();
-    vk::DestroyDescriptorSetLayout(m_device->device(), ds_layout, nullptr);
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkCmdPushDescriptorSetWithTemplateKHR-pData-01686");
-    vk::CmdPushDescriptorSetWithTemplateKHR(m_commandBuffer->handle(), update_template, pipeline_layout, 0, &update_template_data);
-    m_errorMonitor->VerifyFound();
-    m_commandBuffer->end();
-
-    vk::DestroyDescriptorUpdateTemplateKHR(m_device->device(), update_template, nullptr);
-    vk::DestroyPipelineLayout(m_device->device(), pipeline_layout, nullptr);
-}
-
 TEST_F(NegativeDescriptors, CreateDescriptorPool) {
     TEST_DESCRIPTION("Attempt to create descriptor pool with invalid parameters");
 
@@ -2790,300 +2550,6 @@ TEST_F(NegativeDescriptors, DuplicateDescriptorBinding) {
     m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkDescriptorSetLayoutCreateInfo-binding-00279");
     vk::CreateDescriptorSetLayout(m_device->device(), &ds_layout_ci, NULL, &ds_layout);
     m_errorMonitor->VerifyFound();
-}
-
-TEST_F(NegativeDescriptors, IPushDescriptorSetLayout) {
-    TEST_DESCRIPTION("Create a push descriptor set layout with invalid bindings.");
-
-    AddRequiredExtensions(VK_KHR_PUSH_DESCRIPTOR_EXTENSION_NAME);
-
-    ASSERT_NO_FATAL_FAILURE(InitFramework(m_errorMonitor));
-    if (!AreRequiredExtensionsEnabled()) {
-        GTEST_SKIP() << RequiredExtensionsNotSupported() << " not supported";
-    }
-    ASSERT_NO_FATAL_FAILURE(InitState());
-
-    // Get the push descriptor limits
-    auto push_descriptor_prop = LvlInitStruct<VkPhysicalDevicePushDescriptorPropertiesKHR>();
-    GetPhysicalDeviceProperties2(push_descriptor_prop);
-
-    VkDescriptorSetLayoutBinding binding = {0, VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC, 1, VK_SHADER_STAGE_FRAGMENT_BIT, nullptr};
-
-    auto ds_layout_ci = LvlInitStruct<VkDescriptorSetLayoutCreateInfo>();
-    ds_layout_ci.flags = VK_DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR;
-    ds_layout_ci.bindingCount = 1;
-    ds_layout_ci.pBindings = &binding;
-
-    // Note that as binding is referenced in ds_layout_ci, it is effectively in the closure by reference as well.
-    auto test_create_ds_layout = [&ds_layout_ci, this](std::string error) {
-        VkDescriptorSetLayout ds_layout = VK_NULL_HANDLE;
-        m_errorMonitor->SetDesiredFailureMsg(kErrorBit, error);
-        vk::CreateDescriptorSetLayout(m_device->handle(), &ds_layout_ci, nullptr, &ds_layout);
-        m_errorMonitor->VerifyFound();
-    };
-
-    // Starting with the initial VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC type set above..
-    test_create_ds_layout("VUID-VkDescriptorSetLayoutCreateInfo-flags-00280");
-
-    binding.descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC;
-    test_create_ds_layout(
-        "VUID-VkDescriptorSetLayoutCreateInfo-flags-00280");  // This is the same VUID as above, just a second error condition.
-
-    if (!(push_descriptor_prop.maxPushDescriptors == std::numeric_limits<uint32_t>::max())) {
-        binding.descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-        binding.descriptorCount = push_descriptor_prop.maxPushDescriptors + 1;
-        test_create_ds_layout("VUID-VkDescriptorSetLayoutCreateInfo-flags-00281");
-    } else {
-        printf("maxPushDescriptors is set to maximum unit32_t value, skipping 'out of range test'.\n");
-    }
-}
-
-TEST_F(NegativeDescriptors, PushDescriptorImageLayout) {
-    TEST_DESCRIPTION("Use a push descriptor with a mismatched image layout.");
-
-    AddRequiredExtensions(VK_KHR_PUSH_DESCRIPTOR_EXTENSION_NAME);
-    ASSERT_NO_FATAL_FAILURE(InitFramework(m_errorMonitor));
-    if (!AreRequiredExtensionsEnabled()) {
-        GTEST_SKIP() << RequiredExtensionsNotSupported() << " not supported";
-    }
-
-    ASSERT_NO_FATAL_FAILURE(InitState(nullptr, nullptr, VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT));
-    ASSERT_NO_FATAL_FAILURE(InitViewport());
-    ASSERT_NO_FATAL_FAILURE(InitRenderTarget());
-
-    VkDescriptorSetLayoutBinding dsl_binding = {};
-    dsl_binding.binding = 0;
-    dsl_binding.descriptorType = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
-    dsl_binding.descriptorCount = 1;
-    dsl_binding.stageFlags = VK_SHADER_STAGE_FRAGMENT_BIT;
-    dsl_binding.pImmutableSamplers = NULL;
-
-    const VkDescriptorSetLayoutObj ds_layout(m_device, {dsl_binding}, VK_DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR);
-    auto pipeline_layout = VkPipelineLayoutObj(m_device, {&ds_layout});
-
-    char const *fsSource = R"glsl(
-        #version 450
-        layout(set=0, binding=0) uniform sampler2D tex;
-        layout(location=0) out vec4 color;
-        void main(){
-           color = textureLod(tex, vec2(0.5, 0.5), 0.0);
-        }
-    )glsl";
-    VkShaderObj vs(this, kVertexMinimalGlsl, VK_SHADER_STAGE_VERTEX_BIT);
-    VkShaderObj fs(this, fsSource, VK_SHADER_STAGE_FRAGMENT_BIT);
-    VkPipelineObj pipe(m_device);
-    pipe.SetViewport(m_viewports);
-    pipe.SetScissor(m_scissors);
-    pipe.AddDefaultColorAttachment();
-    pipe.AddShader(&vs);
-    pipe.AddShader(&fs);
-    pipe.MakeDynamic(VK_DYNAMIC_STATE_VIEWPORT);
-    pipe.MakeDynamic(VK_DYNAMIC_STATE_SCISSOR);
-    pipe.CreateVKPipeline(pipeline_layout.handle(), m_renderPass);
-
-    vk_testing::Sampler sampler(*m_device, SafeSaneSamplerCreateInfo());
-
-    VkImageObj image(m_device);
-    image.Init(32, 32, 1, VK_FORMAT_B8G8R8A8_UNORM, VK_IMAGE_USAGE_SAMPLED_BIT | VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT);
-    VkImageView image_view = image.targetView(VK_FORMAT_B8G8R8A8_UNORM);
-    image.SetLayout(VK_IMAGE_ASPECT_COLOR_BIT, VK_IMAGE_LAYOUT_GENERAL);
-
-    VkDescriptorImageInfo img_info = {};
-    img_info.sampler = sampler.handle();
-    img_info.imageView = image_view;
-    img_info.imageLayout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-    auto descriptor_write = LvlInitStruct<VkWriteDescriptorSet>();
-    descriptor_write.dstSet = 0;
-    descriptor_write.descriptorCount = 1;
-    descriptor_write.descriptorType = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
-    descriptor_write.pImageInfo = &img_info;
-    descriptor_write.dstArrayElement = 0;
-    descriptor_write.dstBinding = 0;
-
-    for (uint32_t i = 0; i < 2; i++) {
-        m_commandBuffer->begin();
-        if (i == 1) {
-            // Test path where image layout in command buffer is known at draw time
-            image.ImageMemoryBarrier(m_commandBuffer, VK_IMAGE_ASPECT_COLOR_BIT, VK_ACCESS_SHADER_WRITE_BIT,
-                                     VK_ACCESS_SHADER_WRITE_BIT, VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL,
-                                     VK_PIPELINE_STAGE_FRAGMENT_SHADER_BIT, VK_PIPELINE_STAGE_FRAGMENT_SHADER_BIT);
-        }
-        m_commandBuffer->BeginRenderPass(m_renderPassBeginInfo);
-        vk::CmdBindPipeline(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipe.handle());
-        vk::CmdPushDescriptorSetKHR(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipeline_layout.handle(), 0, 1,
-                                    &descriptor_write);
-        vk::CmdSetViewport(m_commandBuffer->handle(), 0, 1, &m_viewports[0]);
-        vk::CmdSetScissor(m_commandBuffer->handle(), 0, 1, &m_scissors[0]);
-
-        if (i == 1) {
-            // Test path where image layout in command buffer is known at draw time
-            m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkCmdDraw-None-08114");
-            m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkDescriptorImageInfo-imageLayout-00344");
-            m_commandBuffer->Draw(1, 1, 0, 0);
-            m_errorMonitor->VerifyFound();
-            break;
-        }
-        m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "UNASSIGNED-CoreValidation-DrawState-InvalidImageLayout");
-        m_commandBuffer->Draw(1, 1, 0, 0);
-        m_commandBuffer->EndRenderPass();
-        m_commandBuffer->end();
-
-        m_commandBuffer->QueueCommandBuffer(false);
-        m_errorMonitor->VerifyFound();
-    }
-}
-
-TEST_F(NegativeDescriptors, PushDescriptorSetLayoutWithoutExtension) {
-    TEST_DESCRIPTION("Create a push descriptor set layout without loading the needed extension.");
-    ASSERT_NO_FATAL_FAILURE(Init());
-
-    VkDescriptorSetLayoutBinding binding = {0, VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER, 1, VK_SHADER_STAGE_FRAGMENT_BIT, nullptr};
-
-    auto ds_layout_ci = LvlInitStruct<VkDescriptorSetLayoutCreateInfo>();
-    ds_layout_ci.flags = VK_DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR;
-    ds_layout_ci.bindingCount = 1;
-    ds_layout_ci.pBindings = &binding;
-
-    VkDescriptorSetLayout ds_layout = VK_NULL_HANDLE;
-
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "UNASSIGNED-CoreValidation-DrawState-ExtensionNotEnabled");
-    vk::CreateDescriptorSetLayout(m_device->handle(), &ds_layout_ci, nullptr, &ds_layout);
-    m_errorMonitor->VerifyFound();
-}
-
-TEST_F(NegativeDescriptors, AllocatePushDescriptorSet) {
-    TEST_DESCRIPTION("Attempt to allocate a push descriptor set.");
-    AddRequiredExtensions(VK_KHR_PUSH_DESCRIPTOR_EXTENSION_NAME);
-
-    ASSERT_NO_FATAL_FAILURE(InitFramework(m_errorMonitor));
-    if (!AreRequiredExtensionsEnabled()) {
-        GTEST_SKIP() << RequiredExtensionsNotSupported() << " not supported";
-    }
-    ASSERT_NO_FATAL_FAILURE(InitState());
-
-    VkDescriptorSetLayoutBinding binding = {0, VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER, 1, VK_SHADER_STAGE_FRAGMENT_BIT, nullptr};
-    auto ds_layout_ci = LvlInitStruct<VkDescriptorSetLayoutCreateInfo>();
-    ds_layout_ci.flags = VK_DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR;
-    ds_layout_ci.bindingCount = 1;
-    ds_layout_ci.pBindings = &binding;
-    vk_testing::DescriptorSetLayout ds_layout(*m_device, ds_layout_ci);
-
-    VkDescriptorPoolSize pool_size = {binding.descriptorType, binding.descriptorCount};
-    auto dspci = LvlInitStruct<VkDescriptorPoolCreateInfo>();
-    dspci.poolSizeCount = 1;
-    dspci.pPoolSizes = &pool_size;
-    dspci.maxSets = 1;
-    vk_testing::DescriptorPool pool(*m_device, dspci);
-
-    auto ds_alloc_info = LvlInitStruct<VkDescriptorSetAllocateInfo>();
-    ds_alloc_info.descriptorPool = pool.handle();
-    ds_alloc_info.descriptorSetCount = 1;
-    ds_alloc_info.pSetLayouts = &ds_layout.handle();
-
-    VkDescriptorSet ds = VK_NULL_HANDLE;
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkDescriptorSetAllocateInfo-pSetLayouts-00308");
-    vk::AllocateDescriptorSets(m_device->handle(), &ds_alloc_info, &ds);
-    m_errorMonitor->VerifyFound();
-}
-
-TEST_F(NegativeDescriptors, CreateDescriptorUpdateTemplate) {
-    TEST_DESCRIPTION("Verify error messages for invalid vkCreateDescriptorUpdateTemplate calls.");
-
-#ifdef __ANDROID__
-    GTEST_SKIP() << "Skipped on Android pending further investigation.";
-#endif
-
-    AddRequiredExtensions(VK_KHR_PUSH_DESCRIPTOR_EXTENSION_NAME);
-    AddRequiredExtensions(VK_KHR_DESCRIPTOR_UPDATE_TEMPLATE_EXTENSION_NAME);
-    ASSERT_NO_FATAL_FAILURE(InitFramework());
-    if (!AreRequiredExtensionsEnabled()) {
-        GTEST_SKIP() << RequiredExtensionsNotSupported() << " not supported.";
-    }
-    ASSERT_NO_FATAL_FAILURE(InitState());
-
-    VkDescriptorSetLayoutBinding dsl_binding = {};
-    dsl_binding.binding = 0;
-    dsl_binding.descriptorType = VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER;
-    dsl_binding.descriptorCount = 1;
-    dsl_binding.stageFlags = VK_SHADER_STAGE_ALL;
-    dsl_binding.pImmutableSamplers = NULL;
-
-    const VkDescriptorSetLayoutObj ds_layout_ub(m_device, {dsl_binding});
-    const VkDescriptorSetLayoutObj ds_layout_ub1(m_device, {dsl_binding});
-    const VkDescriptorSetLayoutObj ds_layout_ub_push(m_device, {dsl_binding},
-                                                     VK_DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR);
-    const VkPipelineLayoutObj pipeline_layout(m_device, {{&ds_layout_ub, &ds_layout_ub1, &ds_layout_ub_push}});
-
-    constexpr uint64_t badhandle = 0xcadecade;
-    VkDescriptorUpdateTemplateEntry entries = {0, 0, 1, VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER, 0, sizeof(VkBuffer)};
-    auto create_info = LvlInitStruct<VkDescriptorUpdateTemplateCreateInfo>();
-    create_info.flags = 0;
-    create_info.descriptorUpdateEntryCount = 1;
-    create_info.pDescriptorUpdateEntries = &entries;
-
-    auto do_test = [&](std::string err) {
-        VkDescriptorUpdateTemplateKHR dut = VK_NULL_HANDLE;
-        m_errorMonitor->SetDesiredFailureMsg(kErrorBit, err);
-        if (VK_SUCCESS == vk::CreateDescriptorUpdateTemplateKHR(m_device->handle(), &create_info, nullptr, &dut)) {
-            vk::DestroyDescriptorUpdateTemplateKHR(m_device->handle(), dut, nullptr);
-        }
-        m_errorMonitor->VerifyFound();
-    };
-
-    // Descriptor set type template
-    create_info.templateType = VK_DESCRIPTOR_UPDATE_TEMPLATE_TYPE_DESCRIPTOR_SET;
-    // descriptorSetLayout is NULL
-    do_test("VUID-VkDescriptorUpdateTemplateCreateInfo-templateType-00350");
-
-    // Bad pipelineLayout handle, to be ignored if template type is DESCRIPTOR_SET
-    {
-        create_info.pipelineLayout = CastFromUint64<VkPipelineLayout>(badhandle);
-        create_info.descriptorSetLayout = ds_layout_ub.handle();
-        VkDescriptorUpdateTemplateKHR dut = VK_NULL_HANDLE;
-        if (VK_SUCCESS == vk::CreateDescriptorUpdateTemplateKHR(m_device->handle(), &create_info, nullptr, &dut)) {
-            vk::DestroyDescriptorUpdateTemplateKHR(m_device->handle(), dut, nullptr);
-        }
-    }
-
-    create_info.templateType = VK_DESCRIPTOR_UPDATE_TEMPLATE_TYPE_PUSH_DESCRIPTORS_KHR;
-    // Bad pipelineLayout handle
-    do_test("VUID-VkDescriptorUpdateTemplateCreateInfo-templateType-00352");
-
-    create_info.pipelineBindPoint = VK_PIPELINE_BIND_POINT_COMPUTE;
-    create_info.pipelineLayout = pipeline_layout.handle();
-    create_info.set = 2;
-
-    // Bad bindpoint -- force fuzz the bind point
-    memset(&create_info.pipelineBindPoint, 0xFE, sizeof(create_info.pipelineBindPoint));
-    do_test("VUID-VkDescriptorUpdateTemplateCreateInfo-templateType-00351");
-    create_info.pipelineBindPoint = VK_PIPELINE_BIND_POINT_COMPUTE;
-
-    // Bad pipeline layout
-    create_info.pipelineLayout = VK_NULL_HANDLE;
-    do_test("VUID-VkDescriptorUpdateTemplateCreateInfo-templateType-00352");
-    create_info.pipelineLayout = pipeline_layout.handle();
-
-    // Wrong set #
-    create_info.set = 0;
-    do_test("VUID-VkDescriptorUpdateTemplateCreateInfo-templateType-00353");
-
-    // Invalid set #
-    create_info.set = 42;
-    do_test("VUID-VkDescriptorUpdateTemplateCreateInfo-templateType-00353");
-
-    // Bad descriptorSetLayout handle, to be ignored if templateType is PUSH_DESCRIPTORS
-    // NOTE: AMD's Windows proprietary driver doesn't seem to ignore this handle
-    create_info.set = 2;
-    create_info.descriptorSetLayout = CastFromUint64<VkDescriptorSetLayout>(badhandle);
-    if (!IsDriver(VK_DRIVER_ID_AMD_PROPRIETARY)) {
-        VkDescriptorUpdateTemplateKHR dut = VK_NULL_HANDLE;
-        if (VK_SUCCESS == vk::CreateDescriptorUpdateTemplateKHR(m_device->handle(), &create_info, nullptr, &dut)) {
-            vk::DestroyDescriptorUpdateTemplateKHR(m_device->handle(), dut, nullptr);
-        }
-    }
-    // Bad descriptorSetLayout handle
-    create_info.templateType = VK_DESCRIPTOR_UPDATE_TEMPLATE_TYPE_DESCRIPTOR_SET;
-    do_test("VUID-VkDescriptorUpdateTemplateCreateInfo-templateType-00350");
 }
 
 TEST_F(NegativeDescriptors, InlineUniformBlockEXT) {
@@ -4278,9 +3744,7 @@ TEST_F(NegativeDescriptors, MutableDescriptorSetLayout) {
 
     AddRequiredExtensions(VK_EXT_MUTABLE_DESCRIPTOR_TYPE_EXTENSION_NAME);
     AddRequiredExtensions(VK_EXT_DESCRIPTOR_INDEXING_EXTENSION_NAME);
-    AddOptionalExtensions(VK_KHR_PUSH_DESCRIPTOR_EXTENSION_NAME);
     ASSERT_NO_FATAL_FAILURE(InitFramework(m_errorMonitor));
-    const bool push_descriptors = IsExtensionsEnabled(VK_KHR_PUSH_DESCRIPTOR_EXTENSION_NAME);
     if (!AreRequiredExtensionsEnabled()) {
         GTEST_SKIP() << RequiredExtensionsNotSupported() << " not supported";
     }
@@ -4325,29 +3789,6 @@ TEST_F(NegativeDescriptors, MutableDescriptorSetLayout) {
 
     binding.pImmutableSamplers = nullptr;
     sampler.destroy();
-
-    if (push_descriptors) {
-        list.descriptorTypeCount = 0;
-        binding.descriptorType = VK_DESCRIPTOR_TYPE_SAMPLER;
-        ds_layout_ci.flags =
-            VK_DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR | VK_DESCRIPTOR_SET_LAYOUT_CREATE_HOST_ONLY_POOL_BIT_EXT;
-        m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkDescriptorSetLayoutCreateInfo-flags-04590");
-        vk::CreateDescriptorSetLayout(m_device->device(), &ds_layout_ci, nullptr, &ds_layout);
-        m_errorMonitor->VerifyFound();
-
-        ds_layout_ci.flags =
-            VK_DESCRIPTOR_SET_LAYOUT_CREATE_UPDATE_AFTER_BIND_POOL_BIT | VK_DESCRIPTOR_SET_LAYOUT_CREATE_HOST_ONLY_POOL_BIT_EXT;
-        m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkDescriptorSetLayoutCreateInfo-flags-04592");
-        vk::CreateDescriptorSetLayout(m_device->device(), &ds_layout_ci, nullptr, &ds_layout);
-        m_errorMonitor->VerifyFound();
-
-        list.descriptorTypeCount = 2;
-        binding.descriptorType = VK_DESCRIPTOR_TYPE_MUTABLE_EXT;
-        ds_layout_ci.flags = VK_DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR;
-        m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkDescriptorSetLayoutCreateInfo-flags-04591");
-        vk::CreateDescriptorSetLayout(m_device->device(), &ds_layout_ci, nullptr, &ds_layout);
-        m_errorMonitor->VerifyFound();
-    }
 }
 
 TEST_F(NegativeDescriptors, MutableDescriptorSetLayoutMissingFeature) {
